requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pymupdf
pandas
openpyxl
requests
python-dotenv
